Generate married couple - v 0.0.35
I suck at this git thing I don't really know when to push or not new things. I don't know version naming either.
Look, I'm learning along the way. We'll see where this goes.
This is kind of the "I need to push that shit to get it over with and start back on clean basis"

So here we have created people 0.0.1, then added the sex 0.0.2, now we have married couples 0.0.3 and the 5 is because I started thinking about the kids so yeah there's that.
Anyway. Bye.