Generate progeny - V.0.0.4
The script now test if the couple have had time to have kids,and if so, how much.
Switched to a grid layout to have generations on differents lines. Will probably need something else later when I'll stop generating only one progeny every generation.
Next step is to add the links between the kids and the parents.

Also when there are too many kids, I have to zoom out to see, else one of them is eaten by the left side of the windows weirdly enough.
Need to investigate the CSS of that.