Quick Update - Shifting gears
No code updates in this one. Just rambling comment in the script.js about how to go about it.
CSS is going to be a nightmare and probably will feel very clunky.
I think SVG is the real thing that willmake it possible.
It's not impossible also that later, neo4j would meddle with this.